+++ table/block_based_table_builder.cc
@@ -276,7 +276,7 @@ struct BlockBasedTableBuilder::Rep {
   uint32_t column_family_id;
   const std::string& column_family_name;
   uint64_t creation_time = 0;
-  uint64_t earliest_key_time = 0;
+  uint64_t earliest_key_time = std::numeric_limits<uint64_t>::max();
 
   std::vector<std::unique_ptr<IntTblPropCollector>> table_properties_collectors;